Cleanup

+++ GalaxyWars/Program.cs
@@ -29,7 +29,7 @@ namespace GalaxyWars
 
         while (spacePope.population > 0 && nasa.population > 0 && stormTroopers.population > 0)
         {
-                Console.WriteLine("It is year {0}.", year);
+            Console.WriteLine("It is year {0}.", year);
             if (year > 1) // if at least one year has passed, grant bonus pop to science type
             {
                 foreach (AlienSpecies species in speciesList)
@@ -55,42 +55,51 @@ namespace GalaxyWars
                 newKey = ruleToChange.Value; // create references for new key value pair
                 newValue = ruleToChange.Key;
 
+                Console.WriteLine("In the last era, {0} defeated {1}.", ruleToChange.Key, ruleToChange.Value);
+                Console.WriteLine("Now, the tables are turned. {0} will defeat {1}.", newKey, newValue);
+
                 battleRules.Remove(ruleToChange); // remove old rule
                 battleRules.Add(new KeyValuePair<string, string>(newKey, newValue)); // add new reversed rule
             }
 
             foreach (AlienSpecies currAttacker in speciesList) // let each species attack once
             {
+                Console.WriteLine("{0} is attacking!", currAttacker.name);
                 foreach (AlienSpecies currDefender in speciesList) // let each species defend once against each attacker
-                {
-                    if (currAttacker.name != currDefender.name) // apply bonus attack from battle rules
-                    {
-                        var whoAttackerBeats = new List<string> { };
-                        foreach (KeyValuePair<string, string> rule in battleRules) // loop through rules
+                    {   
+                        if (currAttacker.name != currDefender.name) // apply bonus attack from battle rules
                         {
-                            if (rule.Key == currAttacker.affiliation) // add each type attacker defeats to list
+                            Console.WriteLine("{0}'s current opponent is {1}.", currAttacker.name, currDefender.name);
+                            var whoAttackerBeats = new List<string> { };
+                            foreach (KeyValuePair<string, string> rule in battleRules) // loop through rules
                             {
-                                whoAttackerBeats.Add(rule.Value);
+                                if (rule.Key == currAttacker.affiliation) // add each type attacker defeats to list
+                                {
+                                    whoAttackerBeats.Add(rule.Value);
+                                }
+                            }
+                            foreach (string defenderType in whoAttackerBeats)
+                            {
+                                if (currDefender.affiliation == defenderType)
+                                {
+                                    currDefender.population -= Math.Floor(currDefender.population * .02);
+                                    Console.WriteLine("{0} attacks {1} and wins. Their population is reduced to {2}.", currAttacker.name, currDefender.name, currDefender.population);
+                                }
+                                else
+                                {
+                                    Console.WriteLine("{0}'s attacks are not effective against {1}.", currAttacker.name, currDefender.name);
+                                }
+                                // make sure no one is dead
                             }
                         }
-                        foreach (string defenderType in whoAttackerBeats)
-                        {
-                            if (currDefender.affiliation == defenderType)
-                            {
-                                currDefender.population -= Math.Floor(currDefender.population * .02);
-                                Console.WriteLine("{0} attacks {1}. Their population is reduced to {2}.", currAttacker.name, currDefender.name, currDefender.population);
-                            }
-                            // make sure no one is dead
-                        }
-                    }
 
-                    if (currAttacker.affiliation == "Warrior") // add warrior type bonus
+                    if (currAttacker.affiliation == "Warrior" && currDefender.affiliation != "Warrior") // add warrior type bonus
                     {
                         currDefender.population -= 10000;
                         Console.WriteLine("{0}, the warriors, kill extra 10000. {1}'s population is now {2}.", currAttacker.name, currDefender.name, currDefender.population);
                         // make sure no one is dead
                     }
-                    if (currAttacker.affiliation == "Religion")
+                    if (currAttacker.affiliation == "Religion" && currDefender.affiliation != "Religion")
                     {
                         double bonus = Math.Floor(currDefender.population * .01);
                         currAttacker.population += bonus; // religious attacker gains some population from defender
@@ -101,11 +110,11 @@ namespace GalaxyWars
                 }
             }
 
-            // base population decline
+                // base population decline
+            Console.WriteLine("As a result of war, all populations lose 20000 points.");
             foreach (AlienSpecies species in speciesList)
             {
                 species.population -= 20000;
-                Console.WriteLine("As a result of war, all populations lose 20000 points.");
                     if (species.population <= 0)
                     {
                     Console.WriteLine("***{0} is dead! GAME OVER***", species.name);
